+++ src/components/scenes/GuiPluginListScene.js
@@ -1,5 +1,4 @@
 // @flow
-
 import AsyncStorage from '@react-native-community/async-storage'
 import { asObject, asString } from 'cleaners'
 import { createInputModal } from 'edge-components'
@@ -12,14 +11,17 @@ import { connect } from 'react-redux'
 
 import { updateOneSetting } from '../../actions/SettingsActions.js'
 import paymentTypeLogoApplePay from '../../assets/images/paymentTypes/paymentTypeLogoApplePay.png'
+import paymentTypeLogoAuspost from '../../assets/images/paymentTypes/paymentTypeLogoAuspost.png'
 import paymentTypeLogoBankgirot from '../../assets/images/paymentTypes/paymentTypeLogoBankgirot.png'
 import paymentTypeLogoBankTransfer from '../../assets/images/paymentTypes/paymentTypeLogoBankTransfer.png'
+import paymentTypeLogoBpay from '../../assets/images/paymentTypes/paymentTypeLogoBpay.png'
 import paymentTypeLogoCash from '../../assets/images/paymentTypes/paymentTypeLogoCash.png'
 import paymentTypeLogoCreditCard from '../../assets/images/paymentTypes/paymentTypeLogoCreditCard.png'
 import paymentTypeLogoFasterPayments from '../../assets/images/paymentTypes/paymentTypeLogoFasterPayments.png'
 import paymentTypeLogoGiftCard from '../../assets/images/paymentTypes/paymentTypeLogoGiftCard.png'
 import paymentTypeLogoIdeal from '../../assets/images/paymentTypes/paymentTypeLogoIdeal.png'
 import paymentTypeLogoNewsagent from '../../assets/images/paymentTypes/paymentTypeLogoNewsagent.png'
+import paymentTypeLogoPayid from '../../assets/images/paymentTypes/paymentTypeLogoPayid.png'
 import paymentTypeLogoPoli from '../../assets/images/paymentTypes/paymentTypeLogoPoli.png'
 import paymentTypeLogoSwish from '../../assets/images/paymentTypes/paymentTypeLogoSwish.png'
 import { ARROW_RIGHT, COUNTRY_CODES, FLAG_LOGO_URL, PLUGIN_VIEW, PLUGIN_VIEW_LEGACY, SIMPLE_ICONS } from '../../constants/indexConstants.js'
@@ -48,14 +50,17 @@ const sellPluginJson = asGuiPluginJson(require('../../constants/plugins/sellPlug
 
 const paymentTypeLogosById = {
   credit: paymentTypeLogoCreditCard,
+  auspost: paymentTypeLogoAuspost,
   applepay: paymentTypeLogoApplePay,
   bank: paymentTypeLogoBankTransfer,
   bankgirot: paymentTypeLogoBankgirot,
+  bpay: paymentTypeLogoBpay,
   cash: paymentTypeLogoCash,
   fasterPayments: paymentTypeLogoFasterPayments,
   giftcard: paymentTypeLogoGiftCard,
   ideal: paymentTypeLogoIdeal,
   newsagent: paymentTypeLogoNewsagent,
+  payid: paymentTypeLogoPayid,
   poli: paymentTypeLogoPoli,
   swish: paymentTypeLogoSwish
 }